+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## 0.1.1
+## 0.1.3
+
+`2018-9-24`
+Adapta refinements
+Added macOS mojave theme
+Renamed from ayu-one to ayu-adaptive
+
+## 0.1.2
 
 `2018-8-22`